backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
groq==0.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx==0.25.2
python-multipart==0.0.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

